refactor: Tag 컴포넌트 스타일 수정 및 타입 조정
